pyproject.toml
[project]
name = "etl-pdf-extractor"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas Almeida Bezerra",email = "lucasalmeida_097@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "camelot-py (>=1.0.0,<2.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "ghostscript (>=0.7,<0.8)",
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "unicode (>=2.9,<3.0)",
    "dbt-core (>=1.9.4,<2.0.0)",
    "dbt-postgres (>=1.9.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "unidecode (>=1.3.8,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "streamlit (>=1.44.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
black = "^25.1.0"
flake8 = "^7.2.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
skip-string-normalization = false

